requirements.txt
groq
python-dotenv
streamlit
discord.py
reportlab
python-docx
